+++ src/settings/components/form/keymaps-form.scss
@@ -1,0 +1,9 @@
+.form-keymaps-form {
+  column-count: 2;
+  .keymap-fields-group {
+    margin-top: 24px;
+  }
+  .keymap-fields-group:first-of-type {
+    margin-top: 0;
+  }
+}
