+++ src/settings/components/site.scss
@@ -24,14 +24,4 @@
       font-weight: bold;
     }
   }
-
-  .keymap-fields{
-    column-count: 2;
-    .keymap-fields-group {
-      margin-top: 24px;
-    }
-    .keymap-fields-group:first-of-type {
-      margin-top: 0;
-    }
-  }
 }
